Create GuiMASCNN.py
This version adds the argument patience, the program waits until the loss improves after patience is reached. Also, the test module was fixed when a number of nodes (n), where p=d/n and p < t (t = training examples), there was an index error. Now is fixed. The accuracy went down a bit, but it is still acceptable.