Create MASAgentCNN.py
This version performs weight synchronization after every mini-batch. This is to boost accuracy. It calculates the patience "automatically", based on patience_epochs that establishes the stop criteria.
I tested and it improves the accuracy when the epochs is large, for instance for 100 reached like 24.75%, it took 25 minutes to dinish. I am testing if using like 150 epochs would increase the accuracy much more..