Create mainMASCNN.py
This DLMP version selects dynamically the GPU that has enough memory to run the program. 

This change was necessary because the GPUclient aborted the run because the GPU 0 was memory saturated. So now it selects the first  GPU that is able to run the program.

The other modules train, agent and model remains the same as the MASSYNC_FIX of 25-06-23.